+++ Sources/PowerSync/Protocol/db/CrudEntry.swift
@@ -32,20 +32,20 @@ public enum UpdateType: String, Codable, Sendable {
 /// Represents a CRUD (Create, Read, Update, Delete) entry in the system.
 public struct CrudEntry: Sendable {
     /// The unique identifier of the entry.
-    let id: String
-    
+    public let id: String
+
     /// The client ID associated with the entry.
-    let clientId: Int64
-    
+    public let clientId: Int64
+
     /// The type of update operation performed on the entry.
-    let op: UpdateType
+    public let op: UpdateType
 
     /// The name of the table where the entry resides.
-    let table: String
-    
+    public let table: String
+
     /// The transaction ID associated with the entry, if any.
-    let transactionId: Int64?
-    
+    public let transactionId: Int64?
+
     /// User-defined metadata that can be attached to writes.
     ///
     /// This is the value the `_metadata` column had when the write to the database was made,
@@ -53,15 +53,15 @@ public struct CrudEntry: Sendable {
     ///
     /// Note that the `_metadata` column and this field are only available when ``Table/trackMetadata``
     /// is enabled.
-    let metadata: String?
-    
+    public let metadata: String?
+
     /// The operation data associated with the entry, represented as a dictionary of column names to their values.
-    let opDataTyped: JsonParam?
-    
+    public let opDataTyped: JsonParam?
+
     /// The operation data associated with the entry, represented as a dictionary of column names to their values.
     ///
     /// Consider using ``CrudEntry/opDataTyped`` instead, which provides values as typed JSON.
-    var opData: [String: String?]? {
+    public var opData: [String: String?]? {
         get {
             opDataTyped?.mapValues { value in
                 do {
@@ -76,14 +76,14 @@ public struct CrudEntry: Sendable {
     /// Previous values before this change.
     ///
     /// These values can be tracked for `UPDATE` statements when ``Table/trackPreviousValues`` is enabled.
-    let previousValuesTyped: JsonParam?
-    
+    public let previousValuesTyped: JsonParam?
+
     /// Previous values before this change.
     ///
     /// These values can be tracked for `UPDATE` statements when ``Table/trackPreviousValues`` is enabled.
     ///
     /// Consider using ``CrudEntry/previousValuesTyped`` instead, which provides values as typed JSON.
-    var previousValues: [String: String?]? {
+    public var previousValues: [String: String?]? {
         get {
             previousValuesTyped?.mapValues { value in
                 do {
@@ -101,7 +101,7 @@ public struct CrudEntry: Sendable {
         let id = try cursor.getInt64(index: 0)
         let txId = cursor.getInt64Optional(index: 1)
         let data = try cursor.getString(index: 2)
-        
+
         struct CrudJsonEntry: Decodable {
             let id: String
             let op: UpdateType
@@ -110,7 +110,7 @@ public struct CrudEntry: Sendable {
             let metadata: String?
             let old: JsonParam?
         }
-        
+
         let decoder = JSONDecoder()
         var entry: CrudJsonEntry
         do {
@@ -131,7 +131,7 @@ public struct CrudEntry: Sendable {
             nonExhaustive: ()
         )
     }
-    
+
     private static func jsonValueToString(_ value: JsonValue?) throws -> String? {
         try value.map { value in
             switch (value) {